use size_t as return type of schedlen(), as there's no error check needed. noted by simonb@netbsd

+++ kame/sys/netinet6/esp.h
@@ -1,4 +1,4 @@
-/*	$KAME: esp.h,v 1.20 2002/08/09 08:49:03 itojun Exp $	*/
+/*	$KAME: esp.h,v 1.21 2003/01/20 00:55:27 itojun Exp $	*/
 
 /*
  * Copyright (C) 1995, 1996, 1997, and 1998 WIDE Project.
@@ -79,7 +79,7 @@ struct esp_algorithm {
 	int (*mature) __P((struct secasvar *));
 	int keymin;	/* in bits */
 	int keymax;	/* in bits */
-	int (*schedlen) __P((const struct esp_algorithm *));
+	size_t (*schedlen) __P((const struct esp_algorithm *));
 	const char *name;
 	int (*ivlen) __P((const struct esp_algorithm *, struct secasvar *));
 	int (*decrypt) __P((struct mbuf *, size_t,
